Fix address[] return types
