+++ web3/utils/abi.py
@@ -2,10 +2,8 @@ import itertools
 
 from eth_abi.abi import (
     process_type,
-    encode_single,
 )
 
-from .encoding import encode_hex
 from .crypto import sha3
 from .string import (
     coerce_args_to_bytes,
@@ -226,85 +224,16 @@ def event_abi_to_log_topic(event_abi):
 
 
 @coerce_return_to_text
-def construct_event_topic_set(event_abi, arguments=None):
-    if arguments is None:
-        arguments = {}
-    if isinstance(arguments, (list, tuple)):
-        if len(arguments) != len(event_abi['inputs']):
-            raise ValueError(
-                "When passing an argument list, the number of arguments must "
-                "match the event constructor."
-            )
-        arguments = {
-            arg['name']: [arg_value]
-            for arg, arg_value
-            in zip(event_abi['inputs'], arguments)
-        }
+def normalize_return_type(data_type, data_value):
+    try:
+        base, sub, arrlist = data_type
+    except ValueError:
+        base, sub, arrlist = process_type(data_type)
 
-    normalized_args = {
-        key: value if is_array(value) else [value]
-        for key, value in arguments.items()
-    }
-
-    event_topic = event_abi_to_log_topic(event_abi)
-    indexed_args = get_indexed_event_inputs(event_abi)
-    zipped_abi_and_args = [
-        (arg, normalized_args.get(arg['name'], [None]))
-        for arg in indexed_args
-    ]
-    encoded_args = [
-        [
-            None if option is None else encode_hex(encode_single(arg['type'], option))
-            for option in arg_options]
-        for arg, arg_options in zipped_abi_and_args
-    ]
-
-    topics = [
-        [event_topic] + list(permutation)
-        if any(value is not None for value in permutation)
-        else [event_topic]
-        for permutation in itertools.product(*encoded_args)
-    ]
-    return topics
-
-
-@coerce_return_to_text
-def construct_event_data_set(event_abi, arguments=None):
-    if arguments is None:
-        arguments = {}
-    if isinstance(arguments, (list, tuple)):
-        if len(arguments) != len(event_abi['inputs']):
-            raise ValueError(
-                "When passing an argument list, the number of arguments must "
-                "match the event constructor."
-            )
-        arguments = {
-            arg['name']: [arg_value]
-            for arg, arg_value
-            in zip(event_abi['inputs'], arguments)
-        }
-
-    normalized_args = {
-        key: value if is_array(value) else [value]
-        for key, value in arguments.items()
-    }
-
-    indexed_args = exclude_indexed_event_inputs(event_abi)
-    zipped_abi_and_args = [
-        (arg, normalized_args.get(arg['name'], [None]))
-        for arg in indexed_args
-    ]
-    encoded_args = [
-        [
-            None if option is None else encode_hex(encode_single(arg['type'], option))
-            for option in arg_options]
-        for arg, arg_options in zipped_abi_and_args
-    ]
-
-    topics = [
-        list(permutation)
-        if any(value is not None for value in permutation)
-        else []
-        for permutation in itertools.product(*encoded_args)
-    ]
-    return topics
+    if arrlist:
+        sub_type = (base, sub, arrlist[:-1])
+        return [normalize_return_type(sub_type, sub_value) for sub_value in data_value]
+    elif base == 'address':
+        return add_0x_prefix(data_value)
+    else:
+        return data_value
